requirements.txt
python-dotenv==1.0.1
mutagen==1.47.0
elevenlabs==1.3.0
pygments==2.18.0
Pillow==10.3.0
pixie-python==4.3.0
black==24.4.2
openai==1.33.0
python-youtube==0.9.4
httplib2==0.22.0
google-api-python-client==2.133.0
google_auth_oauthlib==1.2.0
groq==0.9.0
requests==2.32.3
